Se realizan actualizaciones en registro servicio y reserva articulo

+++ src/main/java/com/uniquindio/bd2/viajesmaravillosos/RegistroServicioController.java
@@ -6,14 +6,14 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
-import javafx.scene.control.TextField;
+import javafx.scene.control.*;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.Optional;
 
 import static com.uniquindio.bd2.viajesmaravillosos.util.Util.asignarColumna;
+import static com.uniquindio.bd2.viajesmaravillosos.util.Util.isConfirmacion;
 
 public class RegistroServicioController {
 
@@ -42,12 +42,30 @@ public class RegistroServicioController {
 
     @FXML
     private void initialize() throws SQLException {
-        asignarColumna(columnaCodigoPaquete, "Codigo Servicio");
-        asignarColumna(columnaCodigoServicio, "Servicio");
-        asignarColumna(columnaServicio, "Codigo Paquete");
+        asignarColumna(columnaCodigoPaquete, "idPaquete");
+        asignarColumna(columnaCodigoServicio, "idServicioPaquete");
+        asignarColumna(columnaServicio, "servicio");
 
         datosTabla = FXCollections.observableArrayList();
         tablaServicio.setItems(datosTabla);
+
+        capturarDatosEnInputs();
+        actualizarTabla();
+    }
+
+    private void capturarDatosEnInputs() {
+        tablaServicio.setOnMouseClicked(event -> {
+            if (event.getClickCount() == 2 && (isConfirmacion("Capturar datos", "¿Estas seguro de sobreescribir los campos?"))){
+                    Servicio servicio = tablaServicio.getSelectionModel().getSelectedItem();
+                    txtServicio.setText(servicio.getServicio());
+                    txtCodigoPaquete.setText(servicio.getIdPaquete());
+                    txtCodigoServicio.setText(servicio.getIdServicioPaquete());
+            }
+        });
+    }
+
+    private void actualizarTabla () throws SQLException {
+        datosTabla.clear();
 
         // Obtener datos de la base de datos y cargar en la tabla
         ResultSet resultado = DataBase.conexion("SELECT * FROM servicio_paquete");
@@ -64,24 +82,29 @@ public class RegistroServicioController {
     }
 
     @FXML
-    void registrarServicio(ActionEvent event) {
+    void registrarServicio(ActionEvent event) throws SQLException {
         DataBase.conexion("INSERT INTO servicio_paquete (id_servicio_paquete, servicio, id_paquete) " +
                 "VALUES (" +
                 "'" + txtCodigoServicio.getText() + "', " +
                 "'" + txtServicio.getText() + "', " +
                 "'" + txtCodigoPaquete.getText() + "')");
+
+        actualizarTabla();
     }
     @FXML
-    void actualizarServicio(ActionEvent event) {
+    void actualizarServicio(ActionEvent event) throws SQLException {
         DataBase.conexion("UPDATE servicio_paquete SET " +
                 "servicio = '" + txtServicio.getText() + "', " +
                 "id_paquete = '" + txtCodigoPaquete.getText() + "' " +
                 "WHERE id_servicio_paquete = '" + txtCodigoServicio.getText() + "'");
+
+        actualizarTabla();
     }
 
     @FXML
     void eliminarServicio(ActionEvent event) {
-        DataBase.conexion("DELETE FROM servicio_paquete WHERE id_servicio_paquete='" + txtCodigoServicio.getText() + "'");
+        if (isConfirmacion("Eliminar registro", "¿Desea eliminar el registro?"))
+            DataBase.conexion("DELETE FROM servicio_paquete WHERE id_servicio_paquete='" + txtCodigoServicio.getText() + "'");
     }
 
     @FXML
